virt plugin: fix build warning
CC       src/virt_la-virt.lo
src/virt.c:417:37: warning: comparison of integers of different signs: 'int' and 'unsigned long' [-Wsign-compare]
  if ((tag_index < 0) || (tag_index >= STATIC_ARRAY_SIZE(tags))) {
                          ~~~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~
1 warning generated.

+++ src/virt.c
@@ -414,7 +414,7 @@ static void memory_stats_submit(gauge_t value, virDomainPtr dom,
                                "minor_fault",    "unused",   "available",
                                "actual_balloon", "rss"};
 
-  if ((tag_index < 0) || (tag_index >= STATIC_ARRAY_SIZE(tags))) {
+  if ((tag_index < 0) || (tag_index >= (int)STATIC_ARRAY_SIZE(tags))) {
     ERROR("virt plugin: Array index out of bounds: tag_index = %d", tag_index);
     return;
   }
